Add folder support
* Reset cursor to idx 0/pg 0 when entering/exiting directory
* Proper path-building w/ '..', hide '..' at root rom dir
* Remove extensions (.vec,.VEC,.bin,.BIN)
* Sorted files/directories
* Modify menu rom to self-identify in order to support directories

Haven't found any bugs yet, and been playing with it for a while.

Since we're sorting the directory and keeping it in-memory, I suppose
there's a limit for the number of files per directory. This was imposed
before via the amount of menu "cart" space that was dedicated to showing
the listings anyways though.

In order to not overuse memory, I've overlaid the directory structure w/
the cartRam space, as their use is mutually exclusive. I've not found any
problem with this, although one could make the argument that the code is
harder to understand.